+++ arm_stdclib.h
@@ -1,6 +1,6 @@
 /*!\file arm_stdclib.h
 ** \author SMFSW
-** \copyright MIT (c) 2017-2021, SMFSW
+** \copyright MIT (c) 2017-2022, SMFSW
 ** \brief ARM common standard c library wrapper macros
 */
 /****************************************************************/